tests: Move ContractSession into its own common module
